+++ src/parser/spells/target.js
@@ -5,7 +5,7 @@ import DICTIONARY from "../../dictionary.js";
  * @param {*} data
  */
 let doesTargetCreature = (data) => {
-  const creature = /You touch a creature|a creature you|creature( that)? you can see|interrupt a creature|would strike a creature|creature of your choice|creature or object within range|cause a creature|creature must be within range/gi;
+  const creature = /You touch a creature|You touch a willing creature|creature you touch|a creature you|creature( that)? you can see|interrupt a creature|would strike a creature|creature of your choice|creature or object within range|cause a creature|creature must be within range/gi;
   const creaturesRange = /(humanoid|monster|creature|target)(s)? (or loose object )?(of your choice )?(that )?(you can see )?within range/gi;
   return data.definition.description.match(creature) || data.definition.description.match(creaturesRange);
 };
@@ -63,8 +63,10 @@ export function getTarget(data) {
 
   switch (data.definition.range.origin) {
     case "Touch":
-      result.units = "touch";
-      if (creatures) result.type = "creature";
+      if (creatures) {
+        result.units = "1";
+        result.type = "creature";
+      }
       break;
     case "Self": {
       const dmgSpell = data.definition.modifiers.some((mod) => mod.type === "damage");