+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Next Up
 
 - On character import update name choice was not respected on prototype token.
+- Fix for somme touch spells not setting target correctly to 1 creature, e.g. Gift of Alacrity.
 
 # 5.1.7